Move WlHeader trait and impl from main.rs to vec_utils.rs
